Cleanup: moved calendar settings to a different file and moved date formats to a IBInspectable variable when apropriate
